Eliminamos firebase.js, lo dividimos en funcionalidades. Agregamos un evento al boton btn-register y un console, pero se esta reiniciando el controlador a la vista de inicio, agregamos window.location.hash a los eventos para que nos genere la ruta en el navegador
